Update build mechanism for genomedata
This brings genomedata up to date with PEP 621. It uses "build" as
the build frontend and "setuptools" as the build backend with
"setuptools-scm" automatically tracking files to include in the
sdist. MANIFEST.in is used to exclude the automatically tracked
files not needed for building and testing from the sdist.

- Updates CI for new build process
- Fixes #59